Move all main actions and types from the user domain, into the main state domain
