+++ src/stores/main/constants.ts
@@ -1,5 +1,26 @@
 export const CLUSTER_SELECT = 'CLUSTER_SELECT';
 
+export const INFO_LOAD_REQUEST = 'INFO_LOAD_REQUEST';
+export const INFO_LOAD_SUCCESS = 'INFO_LOAD_SUCCESS';
+export const INFO_LOAD_ERROR = 'INFO_LOAD_ERROR';
+
 export const GLOBAL_LOAD_REQUEST = 'GLOBAL_LOAD_REQUEST';
 export const GLOBAL_LOAD_SUCCESS = 'GLOBAL_LOAD_SUCCESS';
 export const GLOBAL_LOAD_ERROR = 'GLOBAL_LOAD_ERROR';
+
+export const LOGIN_REQUEST = 'LOGIN_REQUEST';
+export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
+export const LOGIN_ERROR = 'LOGIN_ERROR';
+
+export const LOGOUT_REQUEST = 'LOGOUT_REQUEST';
+export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
+export const LOGOUT_ERROR = 'LOGOUT_ERROR';
+
+export const REFRESH_USER_INFO_REQUEST = 'REFRESH_USER_INFO_REQUEST';
+export const REFRESH_USER_INFO_SUCCESS = 'REFRESH_USER_INFO_SUCCESS';
+export const REFRESH_USER_INFO_ERROR = 'REFRESH_USER_INFO_ERROR';
+
+export const REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST =
+  'REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST';
+export const VERIFY_PASSWORD_RECOVERY_TOKEN = 'VERIFY_PASSWORD_RECOVERY_TOKEN';
+export const SET_NEW_PASSWORD = 'SET_NEW_PASSWORD';
