+++ src/stores/user/constants.ts
@@ -1,15 +1,3 @@
-export const LOGIN_REQUEST = 'LOGIN_REQUEST';
-export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
-export const LOGIN_ERROR = 'LOGIN_ERROR';
-
-export const LOGOUT_REQUEST = 'LOGOUT_REQUEST';
-export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
-export const LOGOUT_ERROR = 'LOGOUT_ERROR';
-
-export const REFRESH_USER_INFO_REQUEST = 'REFRESH_USER_INFO_REQUEST';
-export const REFRESH_USER_INFO_SUCCESS = 'REFRESH_USER_INFO_SUCCESS';
-export const REFRESH_USER_INFO_ERROR = 'REFRESH_USER_INFO_ERROR';
-
 export const USERS_DELETE_REQUEST = 'USERS_DELETE_REQUEST';
 export const USERS_DELETE_SUCCESS = 'USERS_DELETE_SUCCESS';
 export const USERS_DELETE_ERROR = 'USERS_DELETE_ERROR';
@@ -23,15 +11,6 @@ export const USERS_REMOVE_EXPIRATION_REQUEST =
 export const USERS_REMOVE_EXPIRATION_SUCCESS =
   'USERS_REMOVE_EXPIRATION_SUCCESS';
 export const USERS_REMOVE_EXPIRATION_ERROR = 'USERS_REMOVE_EXPIRATION_ERROR';
-
-export const INFO_LOAD_REQUEST = 'INFO_LOAD_REQUEST';
-export const INFO_LOAD_SUCCESS = 'INFO_LOAD_SUCCESS';
-export const INFO_LOAD_ERROR = 'INFO_LOAD_ERROR';
-
-export const REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST =
-  'REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST';
-export const VERIFY_PASSWORD_RECOVERY_TOKEN = 'VERIFY_PASSWORD_RECOVERY_TOKEN';
-export const SET_NEW_PASSWORD = 'SET_NEW_PASSWORD';
 
 export const INVITATION_CREATE_REQUEST = 'INVITATION_CREATE_REQUEST';
 export const INVITATION_CREATE_SUCCESS = 'INVITATION_CREATE_SUCCESS';
