+++ src/stores/user/types.ts
@@ -5,17 +5,6 @@ import {
   INVITATIONS_LOAD_ERROR,
   INVITATIONS_LOAD_REQUEST,
   INVITATIONS_LOAD_SUCCESS,
-  LOGIN_ERROR,
-  LOGIN_REQUEST,
-  LOGIN_SUCCESS,
-  LOGOUT_ERROR,
-  LOGOUT_REQUEST,
-  LOGOUT_SUCCESS,
-  REFRESH_USER_INFO_ERROR,
-  REFRESH_USER_INFO_REQUEST,
-  REFRESH_USER_INFO_SUCCESS,
-  REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST,
-  SET_NEW_PASSWORD,
   USERS_DELETE_ERROR,
   USERS_DELETE_REQUEST,
   USERS_DELETE_SUCCESS,
@@ -25,7 +14,6 @@ import {
   USERS_REMOVE_EXPIRATION_ERROR,
   USERS_REMOVE_EXPIRATION_REQUEST,
   USERS_REMOVE_EXPIRATION_SUCCESS,
-  VERIFY_PASSWORD_RECOVERY_TOKEN,
 } from 'stores/user/constants';
 
 export interface IUserState {
@@ -37,48 +25,6 @@ export interface IUserState {
     isFetching: boolean;
     items: Record<string, IInvitation>;
   };
-}
-
-export interface IUserLoginRequestAction {
-  type: typeof LOGIN_REQUEST;
-  email: string;
-}
-
-export interface IUserLoginSuccessAction {
-  type: typeof LOGIN_SUCCESS;
-  userData: ILoggedInUser;
-}
-
-export interface IUserLoginErrorAction {
-  type: typeof LOGIN_ERROR;
-  errorMessage: string;
-}
-
-export interface IUserLogoutRequestAction {
-  type: typeof LOGOUT_REQUEST;
-}
-
-export interface IUserLogoutSuccessAction {
-  type: typeof LOGOUT_SUCCESS;
-}
-
-export interface IUserLogoutErrorAction {
-  type: typeof LOGOUT_ERROR;
-  errorMessage: string;
-}
-
-export interface IUserRefreshUserInfoRequestAction {
-  type: typeof REFRESH_USER_INFO_REQUEST;
-}
-
-export interface IUserRefreshUserInfoErrorAction {
-  type: typeof REFRESH_USER_INFO_ERROR;
-  error: string;
-}
-
-export interface IUserRefreshUserInfoSuccessAction {
-  type: typeof REFRESH_USER_INFO_SUCCESS;
-  email: string;
 }
 
 export interface IUserLoadRequestAction {
@@ -120,18 +66,6 @@ export interface IUserRemoveExpirationErrorAction {
   type: typeof USERS_REMOVE_EXPIRATION_ERROR;
 }
 
-export interface IUserRequestPasswordRecoveryTokenAction {
-  type: typeof REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST;
-}
-
-export interface IUserVerifyPasswordRecoveryTokenAction {
-  type: typeof VERIFY_PASSWORD_RECOVERY_TOKEN;
-}
-
-export interface IUserSetNewPasswordAction {
-  type: typeof SET_NEW_PASSWORD;
-}
-
 export interface IUserLoadInvitationsRequestAction {
   type: typeof INVITATIONS_LOAD_REQUEST;
 }
@@ -158,15 +92,6 @@ export interface IUserCreateInvitatioErrorAction {
 }
 
 export type UserActions =
-  | IUserLoginRequestAction
-  | IUserLoginSuccessAction
-  | IUserLoginErrorAction
-  | IUserLogoutRequestAction
-  | IUserLogoutSuccessAction
-  | IUserLogoutErrorAction
-  | IUserRefreshUserInfoRequestAction
-  | IUserRefreshUserInfoErrorAction
-  | IUserRefreshUserInfoSuccessAction
   | IUserLoadRequestAction
   | IUserLoadSuccessAction
   | IUserLoadErrorAction
@@ -176,9 +101,6 @@ export type UserActions =
   | IUserDeleteRequestAction
   | IUserDeleteSuccessAction
   | IUserDeleteErrorAction
-  | IUserRequestPasswordRecoveryTokenAction
-  | IUserVerifyPasswordRecoveryTokenAction
-  | IUserSetNewPasswordAction
   | IUserLoadInvitationsRequestAction
   | IUserLoadInvitationsSuccessAction
   | IUserLoadInvitationsErrorAction
